requirements.txt
pillow==10.2.0
python-dotenv==1.0.1
qrcode==7.4.2
validators==0.23.2